reinvent-2019/connected-photo-booth/py_client/requirements.txt
boto3
requests
pillow
piexif
pyexifinfo
pygame
opencv-python
numpy
scikit-image
pyqrcode
pypng
AWSIoTPythonSDK
cups
